requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.3.1+cu121
torchvision==0.18.1+cu121
torchaudio==2.3.1+cu121
pytorch-lightning==2.3.3
peft==0.13.2
transformers==4.45.2
timm>=0.9.16
hydra-core==1.3.2
omegaconf==2.3.0
pillow>=10.0.0
numpy>=1.26.0
matplotlib>=3.10.0
sentencepiece>=0.1.99
safetensors>=0.4.3
